pyproject.toml
[project]
name = "telewish"
version = "1.0.0"
description = "Your minimalistic wish-list with a friendly tg-bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-tailwind>=4.0.1",
    "django>=5.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "django-environ>=0.12.0",
    "django-minio-storage>=0.5.7",
    "python-slugify>=8.0.4",
]

[tool.uv]
default-groups = ["dev", "test", "lint"]

[dependency-groups]
test = [
    "factory-boy>=3.3.3",
    "pytest>=8.3.4",
    "pytest-django>=4.10.0",
]
dev = [
    "cookiecutter>=2.6.0",
    "django-tailwind[reload]>=4.0.1",
]
lint = [
    "pre-commit>=4.2.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "telewish.settings"
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
]

[tool.pyright]
ignore = [
  "tests/factories.py"
]
